web: add verify pages
